update gems & test harness

+++ test/helper.rb
@@ -6,8 +6,9 @@ ROOT_TEST = File.expand_path(File.dirname(__FILE__))
 
 require 'minitest/unit'
 require 'simplecov'
+require 'faker'
 SimpleCov.start do
-  project_name "mHealth - API Server"
+  project_name "API Server"
 
   add_filter "app/app.rb"
   add_filter "/test/"
@@ -18,12 +19,13 @@ SimpleCov.start do
   add_group "Modules", "lib"
   add_group "Routes", "app/routes"
   add_group "Workers", "app/workers"
-
 end
 
 require 'rack/test'
 require 'shoulda'
 require 'pathname'
+require 'mock_redis'
+require 'mocha'
 
 # fire up test server before loading app
 require 'ripple'
@@ -35,10 +37,14 @@ success = false
 # these execute in reverse order on exit
 at_exit { exit! success }
 at_exit { puts 'API Suite Tear Down' }
+at_exit {
+  # Spit out coverage stats.
+  SimpleCov.result.format!
+}
 at_exit do
   unless $! || Test::Unit.run?
     success = Test::Unit::AutoRunner.run
-    Ripple::TestServer.destroy # run after test suite is done
+    Ripple::TestServer.destroy unless Ripple.config[:retain_test_server] = true
     success
   end
 end
@@ -53,16 +59,19 @@ WebMock.disable_net_connect! allow_localhost: true, allow: ["#{Ripple.config[:ho
 
 FIXTURES_PATH = Pathname.new(File.join(ROOT_DIR,'test','support','fixtures'))
 
+# mock redis
+Resque.redis = MockRedis.new
+
 class Test::Unit::TestCase
   include Rack::Test::Methods
 
   def valid_user_details
-    user_data =  {
-      :name => 'Randy Secrist',
-      :gender =>'m',
-      :locale =>'Salt Lake City, UT',
-      :email =>'rsecrist@basho.com',
-      :phone_numbers =>['385-226-1327'],
+    @user_data =  {
+      :name => Faker::Name.name,
+      :gender => 'm',
+      :locale => "#{Faker::Address.city}, #{Faker::Address.state_abbr}",
+      :email => Faker::Internet.email,
+      :phone_numbers =>[Faker::PhoneNumber.phone_number, Faker::PhoneNumber.cell_phone],
       :birthdate => '2013-07-30T11:21:24Z',
       :sources => []
     }
